@@ -34,24 +35,26 @@
     "es"
   ],
   "dependencies": {
-    "@hi-ui/classic-theme": "1.2.1",
-    "@hi-ui/core-css": "^1.0.0",
-    "async-validator": "^1.8.5",
+    "@hi-ui/classic-theme": "1.2.5",
+    "@hi-ui/core-css": "^2.0.0-beta.9",
+    "@mapbox/rehype-prism": "^0.3.1",
+    "async-validator": "^1.11.2",
     "axios": "^0.18.0",
     "classnames": "^2.2.5",
-    "date-fns": "^1.29.0",
+    "clipboard": "^2.0.4",
     "fetch-jsonp": "^1.1.3",
     "github-markdown-css": "^3.0.1",
     "hoist-non-react-statics": "^2.5.0",
     "lodash": "^4.17.10",
-    "react": "^16.3.2",
+    "react": "^16.8.6",
     "react-addons-css-transition-group": "^15.6.2",
     "react-click-outside": "^3.0.1",
     "react-dnd": "^7.4.5",
     "react-dnd-html5-backend": "^7.4.4",
-    "react-dom": "^16.3.2",
+    "react-dom": "^16.8.2",
     "react-lifecycles-compat": "^3.0.4",
-    "react-redux": "^5.0.7",
+    "react-live": "^2.1.1",
+    "react-redux": "^5.1.1",
     "redux": "^4.0.0",
     "redux-thunk": "^2.3.0",
     "shallowequal": "^1.0.2",
@@ -86,34 +92,34 @@
     "html-webpack-plugin": "^3.2.0",
     "husky": "^0.14.3",
     "jest": "^24.8.0",
-    "jest-fetch-mock": "^2.1.1",
+    "jest-fetch-mock": "^2.1.2",
     "jsx-loader": "^0.13.2",
     "lint-staged": "^7.2.2",
     "marked": "^0.4.0",
     "merge2": "^1.2.2",
-    "node-sass": "^4.9.3",
+    "node-sass": "^4.12.0",
     "postcss-loader": "^2.1.5",
-    "prismjs": "^1.14.0",
+    "prismjs": "^1.16.0",
     "raw-loader": "^0.5.1",
     "react-custom-scrollbars": "^4.2.1",
     "react-hot-loader": "4.2.0",
-    "react-test-renderer": "^16.8.2",
-    "rimraf": "^2.6.2",
+    "react-test-renderer": "^16.8.6",
+    "rimraf": "^2.6.3",
     "rucksack-css": "^1.0.2",
     "run-sequence": "^2.2.1",
     "sass-loader": "^7.0.1",
     "standard": "^11.0.1",
     "style-loader": "^0.21.0",
-    "stylelint": "^9.9.0",
-    "stylelint-config-recommended-scss": "^3.2.0",
-    "stylelint-config-standard": "^18.2.0",
-    "stylelint-scss": "^3.4.0",
-    "terser-webpack-plugin": "^1.2.3",
-    "through2": "^2.0.3",
+    "stylelint": "^9.10.1",
+    "stylelint-config-recommended-scss": "^3.3.0",
+    "stylelint-config-standard": "^18.3.0",
+    "stylelint-scss": "^3.7.0",
+    "terser-webpack-plugin": "^1.3.0",
+    "through2": "^2.0.5",
     "url-loader": "^1.0.1",
-    "webpack": "^4.29.0",
-    "webpack-cli": "^3.2.1",
-    "webpack-dev-server": "^3.1.12"
+    "webpack": "^4.32.2",
+    "webpack-cli": "^3.3.2",
+    "webpack-dev-server": "^3.4.1"
   },
   "standard": {
     "env": [
