@@ -1,4 +1,12 @@
-import CheckBox from './CheckBox'
+import Checkbox from './Checkbox'
+import CheckboxLegacy from './checkbox-legacy/index'
+import Group from './Group'
+import SwitchVersion from '../_util/SwitchVersion'
 import './style/index'
 
-export default CheckBox
+const VGroup = SwitchVersion(Group, undefined)
+const VCheckbox = SwitchVersion(Checkbox, CheckboxLegacy)
+
+VCheckbox.Group = VGroup
+
+export default VCheckbox
