@@ -8,25 +8,32 @@ class Collapse extends Component {
   static propTypes = {
     accordion: PropTypes.bool, // 手风琴模式
     activeKey: PropTypes.oneOfType([PropTypes.string, PropTypes.array]),
+    activeId: PropTypes.oneOfType([PropTypes.string, PropTypes.array]),
     onChange: PropTypes.func,
     icon: PropTypes.string,
-    type: PropTypes.string, //
-    arrow: PropTypes.string
+    type: PropTypes.string, // TODO:废弃
+    arrow: PropTypes.oneOf(['left', 'right']), // TODO:废弃，使用 arrowPlacement
+    arrowPlacement: PropTypes.oneOf(['left', 'right']),
+    showArrow: PropTypes.bool
   }
   static defaultProps = {
     prefixCls: 'hi-collapse',
     accordion: false,
     arrow: 'left',
-    type: 'default'
+    arrowPlacement: 'left',
+    type: 'default',
+    showArrow: true
   }
   constructor (props) {
     super(props)
+    const { activeId, activeKey } = this.props
+    const _activeId = activeId || activeKey
     this.state = {
-      activeKey: Array.isArray(this.props.activeKey) ? this.props.activeKey : [this.props.activeKey]
+      activeId: Array.isArray(_activeId) ? _activeId : [_activeId]
     }
   }
   onClickPanel (key) {
-    let activeKey = this.state.activeKey
+    let activeKey = this.state.activeId
     if (this.props.accordion) {
       activeKey = activeKey[0] === key ? [] : [key]
     } else {
@@ -42,20 +49,18 @@ class Collapse extends Component {
     this.setActiveKey(activeKey)
   }
   setActiveKey (activeKey) {
-    // if (!('activeKey' in this.props)) {
-    this.setState({ activeKey })
-    // }
+    this.setState({ activeId: activeKey })
     this.props.onChange(this.props.accordion ? activeKey[0] : activeKey)
   }
 
   renderPanels () {
-    const activeKey = this.state.activeKey
-    const { children, accordion, arrow } = this.props
+    const activeKey = this.state.activeId
+    const { children, accordion, arrow, arrowPlacement, showArrow } = this.props
     const newChildren = []
     Children.forEach(children, (child, index) => {
       if (!child) return
-      const key = child.key || String(index)
-      const { header, disabled } = child.props
+      const key = child.id || child.key || String(index)
+      const { header, disabled, title } = child.props
       let isActive = false
       if (accordion) {
         isActive = activeKey[0] === key
@@ -64,10 +69,11 @@ class Collapse extends Component {
       }
       const props = {
         key,
-        header,
+        header: title || header,
         disabled,
         isActive,
-        arrow,
+        arrow: arrowPlacement !== 'left' ? arrowPlacement : arrow,
+        showArrow,
         children: child.props.children,
         onClickPanel: disabled ? noop : () => this.onClickPanel(key)
       }
@@ -78,11 +84,7 @@ class Collapse extends Component {
   render () {
     const { prefixCls, type } = this.props
     let classnames = classNames(prefixCls, type && `${prefixCls}__${type}`)
-    return (
-      <div className={classnames}>
-        {this.renderPanels()}
-      </div>
-    )
+    return <div className={classnames}>{this.renderPanels()}</div>
   }
 }
 
@@ -92,6 +94,7 @@ class CollapsePanel extends Component {
     disabled: PropTypes.bool,
     isActive: PropTypes.bool,
     arrow: PropTypes.string,
+    showArrow: PropTypes.bool,
     onClickPanel: PropTypes.func
   }
   static defaultProps = {
@@ -99,25 +102,18 @@ class CollapsePanel extends Component {
   }
 
   render () {
-    const {
-      key,
-      arrow,
-      header,
-      disabled,
-      isActive,
-      children,
-      onClickPanel
-    } = this.props
-    let classnames = classNames('collapse-item', { 'collapse-item--show': isActive, 'collapse-item--disabled': disabled })
+    const { key, arrow, header, disabled, isActive, children, onClickPanel, showArrow } = this.props
+    let classnames = classNames('collapse-item', {
+      'collapse-item--show': isActive,
+      'collapse-item--disabled': disabled
+    })
     const collapseIcon = classNames('collapse-item__icon', 'hi-icon', 'icon-right')
     return (
       <div className={classnames}>
         <div className='collapse-item__head' onClick={() => onClickPanel(key)}>
-          { arrow === 'left' && <i className={collapseIcon} /> }
-          <div className='collapse-item__title'>
-            {header}
-          </div>
-          { arrow === 'right' && <i className={collapseIcon} /> }
+          {showArrow && arrow === 'left' && <i className={collapseIcon} />}
+          <div className='collapse-item__title'>{header}</div>
+          {showArrow && arrow === 'right' && <i className={collapseIcon} />}
         </div>
         <div className='collapse-item__content'>{children}</div>
       </div>
