@@ -20,12 +20,13 @@ class Dropdown extends Component {
       PropTypes.arrayOf(PropTypes.string)
     ]),
     onClick: PropTypes.func,
-    list: PropTypes.arrayOf(PropTypes.shape({
+    data: PropTypes.arrayOf(PropTypes.shape({
       title: PropTypes.oneOfType([PropTypes.string, PropTypes.node]),
       onClick: PropTypes.func,
       prefix: PropTypes.oneOfType([PropTypes.string, PropTypes.node]),
       suffix: PropTypes.oneOfType([PropTypes.string, PropTypes.node]),
-      value: PropTypes.oneOfType([PropTypes.string, PropTypes.number])
+      value: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
+      url: PropTypes.string
     })),
     title: PropTypes.oneOfType([PropTypes.string, PropTypes.node]),
     prefix: PropTypes.oneOfType([PropTypes.string, PropTypes.node]),
@@ -36,7 +37,7 @@ class Dropdown extends Component {
     placement: 'top-bottom-start',
     trigger: 'click',
     onClick: () => {},
-    list: []
+    data: []
   }
   componentDidMount () {
     this.registerEvent()
@@ -103,7 +104,7 @@ class Dropdown extends Component {
 
   render () {
     // splitButton，disabled
-    const {list, width, prefix, suffix, theme, title, placement} = this.props
+    const {data, width, prefix, suffix, theme, title, placement} = this.props
     const {visible} = this.state
     const ulCls = classNames('hi-dropdown__menu')
     return (
@@ -118,7 +119,7 @@ class Dropdown extends Component {
         >
           <ul className={ulCls}>
             {
-              list.map((item, index) => {
+              (data).map((item, index) => {
                 if (item.title === '-') {
                   // 分隔线
                   return <li className='hi-dropdown__divider' key={index} />
@@ -130,10 +131,16 @@ class Dropdown extends Component {
                   String(item.title) === String(title) && 'hi-dropdown__item--active'
                 )
 
+                const ItemWrapper = ({href, children}) => {
+                  return href ? <a href={href}>{children}</a> : <React.Fragment>{children}</React.Fragment>
+                }
+
                 return <li className={liCls} key={index} onClick={this.handlerClick.bind(this, item)}>
-                  {(prefix || item.prefix) && <div className='hi-dropdown__item-prefix'>{prefix || item.prefix}</div>}
-                  <div className='hi-dropdown__item-title' title={item.title}>{item.title}</div>
-                  {(suffix || item.suffix) && <div className='hi-dropdown__item-suffix'>{suffix || item.suffix}</div>}
+                  <ItemWrapper href={item.url}>
+                    {(prefix || item.prefix) && <div className='hi-dropdown__item-prefix'>{prefix || item.prefix}</div>}
+                    <div className='hi-dropdown__item-title' title={item.title}>{item.title}</div>
+                    {(suffix || item.suffix) && <div className='hi-dropdown__item-suffix'>{suffix || item.suffix}</div>}
+                  </ItemWrapper>
                 </li>
               })
             }
