@@ -1,4 +1,5 @@
 import Dropdown from './Dropdown'
 import './style/index'
+import { depreactedPropsCompat } from '../_util'
 
-export default Dropdown
+export default depreactedPropsCompat([['data', 'list']])(Dropdown)
