@@ -1,20 +1,14 @@
 import React from 'react'
+import classNames from 'classnames'
 import './style/index'
 
 class Icon extends React.Component {
   render () {
-    const {
-      className,
-      name,
-      style,
-      ...rest
-    } = this.props
+    const { className, name, style } = this.props
 
-    const classes = `hi-icon icon-${name} ${className || ''}`
+    const iconCls = classNames(`hi-icon`, `icon-${name}`, className)
 
-    return (
-      <i className={classes} style={style} {...rest} />
-    )
+    return <i className={iconCls} style={style} />
   }
 }
 
