@@ -32,7 +32,7 @@ export { default as Stepper } from './stepper'
 export { default as confirm } from './confirm'
 export { default as Form } from './form'
 export { default as Menu } from './menu'
-export { default as FormItem } from './form/item'
+export { default as FormItem } from './form/Item'
 export { default as Ficon } from './ficon'
 export { default as Icon } from './icon'
 export { default as Progress } from './progress'
