@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import {createPortal} from 'react-dom'
+import { createPortal } from 'react-dom'
 import PropTypes from 'prop-types'
 import classNames from 'classnames'
 import Button from '../button'
@@ -10,26 +10,32 @@ import Provider from '../context'
 class Modal extends Component {
   static propTypes = {
     title: PropTypes.oneOfType([PropTypes.string, PropTypes.node]),
-    size: PropTypes.oneOf(['large', 'normal', 'small']),
-    show: PropTypes.bool,
+    size: PropTypes.oneOf(['large', 'default', 'small']),
+    show: PropTypes.bool, // TODO: 废弃，使用 visible
+    visible: PropTypes.bool,
     onConfirm: PropTypes.func,
     onCancel: PropTypes.func,
     confirmText: PropTypes.string,
     cancelText: PropTypes.string,
-    backDrop: PropTypes.bool,
+    backDrop: PropTypes.bool, // TODO: 废弃，使用 maskClosable
+    maskClosable: PropTypes.bool,
     confirmType: PropTypes.string,
     cancelType: PropTypes.string,
-    closeBtn: PropTypes.bool
+    closeBtn: PropTypes.bool, // TODO: 废弃，使用 maskClosable
+    closabble: PropTypes.bool
   }
 
   static defaultProps = {
     title: '',
     adaptive: false,
     backDrop: true,
+    maskClosable: true,
     show: false,
+    visible: false,
     closeBtn: true,
+    closable: true,
     destory: false,
-    size: 'normal'
+    size: 'default'
   }
   constructor (props) {
     super(props)
@@ -39,7 +45,7 @@ class Modal extends Component {
   }
   static getDerivedStateFromProps (props, state) {
     // 屏蔽滚动条
-    if (props.show) {
+    if (props.show || props.visible) {
       document.body.style.setProperty('overflow', 'hidden')
     } else {
       document.body.style.removeProperty('overflow')
@@ -66,9 +72,7 @@ class Modal extends Component {
     this._unblock()
   }
   localeDatasProps (key) {
-    const {
-      localeDatas
-    } = this.props
+    const { localeDatas } = this.props
     if (this.props[key]) {
       return this.props[key]
     } else {
@@ -76,11 +80,11 @@ class Modal extends Component {
     }
   }
   renderFooter () {
-    const {footers} = this.props
+    const { footers, footer } = this.props
     const cancelText = this.localeDatasProps('cancelText')
     const confirmText = this.localeDatasProps('confirmText')
-    if (footers) {
-      return footers
+    if (footers || footer) {
+      return footers || footer
     } else {
       return [
         <Button type='default' key={0} onClick={this.handleClose.bind(this)}>
@@ -93,34 +97,54 @@ class Modal extends Component {
     }
   }
   render () {
-    const {width, title, children, backDrop, closeBtn, footers, show, size, className} = this.props
-    let style = width ? {width} : {}
-    let classnames = classNames('hi-modal', show ? '' : 'hi-modal--hide', className)
+    const {
+      width,
+      style = {},
+      title,
+      children,
+      backDrop,
+      maskClosable,
+      closeBtn,
+      closable,
+      footers, // TODO: 废弃，使用 footer 代替
+      footer,
+      show,
+      visible,
+      size,
+      className
+    } = this.props
+    let _style = width ? Object.assign({}, { width }, { ...style }) : style
+    let classnames = classNames('hi-modal', show || visible ? '' : 'hi-modal--hide', className)
     return createPortal(
-      <div className={classnames} >
-        <div className='hi-modal__mask' onClick={() => {
-          backDrop && this.handleClose.apply(this)
-        }} />
-        <div className={classNames('hi-modal__dialog', size ? `hi-modal__dialog--${size}` : '')} style={style}>
-          {
-            (title || closeBtn) && <div className='hi-modal__header'>
+      <div className={classnames}>
+        <div
+          className='hi-modal__mask'
+          onClick={() => {
+            ((backDrop && maskClosable !== false) || (maskClosable && backDrop !== false)) &&
+              this.handleClose.apply(this)
+          }}
+        />
+        <div
+          className={classNames(
+            'hi-modal__dialog',
+            size ? `hi-modal__dialog--${size === 'default' ? 'normal' : size}` : ''
+          )}
+          style={_style}
+        >
+          {(title || (closeBtn && closable !== false) || (closable && closeBtn !== false)) && (
+            <div className='hi-modal__header'>
               <h3 className='hi-modal__title'>{title}</h3>
-              {
-                closeBtn &&
+              {((closeBtn && closable !== false) || (closable && closeBtn !== false)) && (
                 <div className='hi-modal__close' onClick={this.handleClose.bind(this)}>
                   <Icon name='close' />
                 </div>
-              }
+              )}
             </div>
-          }
-          <div className='hi-modal__content'>
-            {children}
-          </div>
-          {
-            (!footers || footers.length > 0) && <div className='hi-modal__footer'>
-              {this.renderFooter()}
-            </div>
-          }
+          )}
+          <div className='hi-modal__content'>{children}</div>
+          {(((!footers || footers.length > 0) && footer !== null) || footer) && (
+            <div className='hi-modal__footer'>{this.renderFooter()}</div>
+          )}
         </div>
       </div>,
       this.node
