@@ -1,2 +1,11 @@
+import Pagination from './Pagination'
 import './style/index'
-export { default } from './Pagination'
+import { depreactedPropsCompat } from '../_util'
+
+export default depreactedPropsCompat([
+  ['type', 'mode'],
+  ['max', 'pageBufferSize'],
+  ['autoHide', 'hideOnSinglePage'],
+  ['showJumper', 'showQuickJumper'],
+  ['onPageSizeChange', 'sizeChangeEvent']
+])(Pagination)
