@@ -1,31 +1,33 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import './style'
-import {BarProgress} from './BarProgress.js'
-import {CircleProgress} from './CircleProgress.js'
-import {DashboardProgress} from './DashboardProgress'
+import { BarProgress } from './BarProgress.js'
+import { CircleProgress } from './CircleProgress.js'
+import { DashboardProgress } from './DashboardProgress'
 
 class Progress extends Component {
   static propTypes = {
-    type: PropTypes.string,
-    status: PropTypes.oneOf(['success', 'warn', 'error', 'primary']),
+    apperance: PropTypes.string,
+    type: PropTypes.oneOf(['success', 'warn', 'error', 'primary']),
     percent: PropTypes.number,
     width: PropTypes.number,
     height: PropTypes.number,
     tooltip: PropTypes.oneOfType([PropTypes.string, PropTypes.node]),
-    text: PropTypes.oneOfType([PropTypes.string, PropTypes.node]),
-    size: PropTypes.oneOf(['big', 'large', 'middle', 'small'])
+    content: PropTypes.oneOfType([PropTypes.string, PropTypes.node]),
+    size: PropTypes.oneOf(['large', 'default', 'small'])
   }
   static defaultProps = {
-    type: 'bar',
-    size: 'large',
+    apperance: 'bar',
+    size: 'default',
     percent: 0,
     radius: 40,
-    status: 'primary'
+    type: 'primary',
+    showInfo: true,
+    placement: 'outside'
   }
 
-  getRenderType (type) {
-    switch (type) {
+  getRenderType (apperance) {
+    switch (apperance) {
       case 'circle':
         return <CircleProgress {...this.props} />
       case 'dashboard':
@@ -37,12 +39,10 @@ class Progress extends Component {
 
   render () {
     let prefix = 'hi-progress'
-    const {className = '', type} = this.props
+    const { className = '', apperance } = this.props
 
     return (
-      <div className={`${prefix}__container ${className}`}>
-        {this.getRenderType(type)}
-      </div>
+      <div className={`${prefix}__container ${className}`}>{this.getRenderType(apperance)}</div>
     )
   }
 }
