@@ -1,4 +1,5 @@
 import './style/index'
 import Rate from './Rate'
+import { depreactedPropsCompat } from '../_util'
 
-export default Rate
+export default depreactedPropsCompat([['clearable', 'allowClear']])(Rate)
