@@ -35,14 +35,18 @@ export default class SelectDropdown extends Component {
       <React.Fragment>
         {mode === 'multiple' && (
           <div className='hi-select__dropdown--item__checkbox'>
-            <Checkbox checked={isSelected} disabled={item.disabled} />
+            <Checkbox checked={isSelected} disabled={item.disabled}>
+              <div className='hi-select__dropdown--item__name'>{item.title}</div>
+            </Checkbox>
           </div>
         )}
-        <div className='hi-select__dropdown--item__name'>{item.name}</div>
+        {mode === 'single' && <div className='hi-select__dropdown--item__name'>{item.title}</div>}
         {mode === 'single' && isSelected && (
-          <div className='hi-select__dropdown--item__check-icon'>
-            <i className='hi-icon icon-check' />
-          </div>
+          <React.Fragment>
+            <div className='hi-select__dropdown--item__check-icon'>
+              <i className='hi-icon icon-check' />
+            </div>
+          </React.Fragment>
         )}
       </React.Fragment>
     )
