@@ -90,7 +90,9 @@ export default class SelectInput extends Component {
       disabled,
       searchable,
       clearable,
-      multipleMode
+      multipleMode,
+      onFocus,
+      onBlur
     } = this.props
     let icon = dropdownShow ? 'up' : 'down'
     let {
@@ -119,7 +121,7 @@ export default class SelectInput extends Component {
           {
             selectedItems.slice(0, showCount).map((item, index) => {
               const _item = <div key={index} className='hi-select__input--item'>
-                <div className='hi-select__input--item__name'>{item.name}</div>
+                <div className='hi-select__input--item__name'>{item.title}</div>
                 <span
                   className='hi-select__input--item__remove'
                   onClick={e => {
@@ -153,6 +155,8 @@ export default class SelectInput extends Component {
                 }}
                 onChange={this.handleKeywordChange.bind(this)}
                 onKeyDown={this.handleKeyDown.bind(this)}
+                onFocus={onFocus.bind(this)}
+                onBlur={onBlur.bind(this)}
               />
             </div>
           }
@@ -172,15 +176,17 @@ export default class SelectInput extends Component {
       dropdownShow,
       disabled,
       searchable,
-      clearable
+      clearable,
+      onFocus,
+      onBlur
     } = this.props
-    placeholder = selectedItems.length > 0 ? selectedItems[0].name : placeholder
+    placeholder = selectedItems.length > 0 ? selectedItems[0].title : placeholder
     let icon = dropdownShow ? 'up' : 'down'
 
     return (
       <div className={classNames('hi-select__input', 'single-value', {disabled})} onClick={this.props.onClick}>
         <div className={classNames('hi-select__input--item', {'hi-select__hide': !(!dropdownShow && selectedItems.length > 0)})}>
-          <div className='hi-select__input--item__name'>{selectedItems[0] && selectedItems[0].name}</div>
+          <div className='hi-select__input--item__name'>{selectedItems[0] && selectedItems[0].title}</div>
         </div>
         {
           (dropdownShow || selectedItems.length === 0) &&
@@ -193,6 +199,8 @@ export default class SelectInput extends Component {
               placeholder={placeholder}
               onChange={this.handleKeywordChange.bind(this)}
               onKeyDown={this.handleKeyDown.bind(this)}
+              onFocus={onFocus.bind(this)}
+              onBlur={onBlur.bind(this)}
               readOnly={disabled || !searchable}
             />
           </div>
