@@ -1,6 +1,9 @@
 import Select from './Select'
 import Option from './Option'
+import SelectLegacy from './select-legacy'
 import './style/index'
+import SwitchVersion from '../_util/SwitchVersion'
 
 Select.Option = Option
-export default Select
+
+export default SwitchVersion(Select, SelectLegacy)
