@@ -2,6 +2,12 @@ import Tabs from './Tabs'
 import TabPane from './TabPane'
 import './style/index'
 
-Tabs.Pane = TabPane
+import SwitchVersion from '../_util/SwitchVersion'
+import TabsLegacy from './tabs-legacy/index'
 
-export default Tabs
+const VTabs = SwitchVersion(Tabs, TabsLegacy)
+const VTabPane = SwitchVersion(TabPane, TabsLegacy.Pane)
+
+VTabs.Pane = VTabPane
+
+export default VTabs
