@@ -6,7 +6,7 @@ import Popper from '../popper'
 import './style/index'
 
 const prefixCls = 'hi-tooltip'
-
+const tooltipInstance = {}
 class Tooltip extends Component {
   static propTypes = {
     defaultVisible: PropTypes.bool,
@@ -28,10 +28,7 @@ class Tooltip extends Component {
 
   render () {
     const { placement, style, className, onClick, title, children } = this.props
-    const eleClass = classNames(
-      `${prefixCls}-base`,
-      placement && `${prefixCls}-${placement}`
-    )
+    const eleClass = classNames(`${prefixCls}-base`, placement && `${prefixCls}-${placement}`)
     const { tooltipShow } = this.state
     return (
       <div
@@ -46,7 +43,7 @@ class Tooltip extends Component {
         onClick={() => {
           onClick && onClick()
         }}
-        ref={(node) => {
+        ref={node => {
           this.tooltipContainer = node
         }}
       >
@@ -66,12 +63,9 @@ class Tooltip extends Component {
   }
 }
 
-function open ({ target, placement = 'top', title }) {
+function deprecatedOpen ({ target, placement = 'top', title }) {
   let mountNode = document.createElement('div')
-  const eleClass = classNames(
-    `${prefixCls}-base`,
-    placement && `${prefixCls}-${placement}`
-  )
+  const eleClass = classNames(`${prefixCls}-base`, placement && `${prefixCls}-${placement}`)
   render(
     <Popper
       className={`${prefixCls}__popper`}
@@ -85,13 +79,47 @@ function open ({ target, placement = 'top', title }) {
     </Popper>,
     mountNode
   )
-  function close () {
+  function deprecatedClose () {
     mountNode && unmountComponentAtNode(mountNode)
     mountNode = undefined
   }
-  return { close }
+  return { close: deprecatedClose }
 }
 
-Tooltip.open = open
+function open (target, { placement = 'top', title, key }) {
+  let mountNode = document.createElement('div')
+  const eleClass = classNames(`${prefixCls}-base`, placement && `${prefixCls}-${placement}`)
+  render(
+    <Popper
+      className={`${prefixCls}__popper`}
+      show
+      attachEle={target}
+      placement={placement}
+      zIndex={1070}
+      width='auto'
+    >
+      <div className={eleClass}>{title}</div>
+    </Popper>,
+    mountNode
+  )
+  tooltipInstance[key] = mountNode
+}
+function close (key) {
+  if (tooltipInstance[key]) {
+    unmountComponentAtNode(tooltipInstance[key])
+    tooltipInstance[key].parentNode &&
+      tooltipInstance[key].parentNode.removeChild(tooltipInstance[key])
+  }
+}
+function openWrapper (target, options) {
+  if (React.isValidElement(React.cloneElement(target))) {
+    open(target, options)
+  } else {
+    return deprecatedOpen(target)
+  }
+}
+
+Tooltip.open = openWrapper
+Tooltip.close = close
 
 export default Tooltip
