@@ -5,16 +5,17 @@ module.exports = {
     template: '布局主题',
     theme: '配色主题',
     changelog: '更新日志',
-    'group-basic': '基本组件',
+    'group-basic': '通用',
     'group-navgation': '导航',
-    'group-form': '表单',
+    'group-form': '数据录入',
     'group-data': '数据展示',
-    'group-tips': '通知/提示',
+    'group-tips': '状态/反馈',
     typography: 'Typography 排版',
     button: 'Button 按钮',
-    dropdown: 'Dropdown 下拉',
+    dropdown: 'Dropdown 下拉菜单',
     pagination: 'Pagination 分页',
     tabs: 'Tabs 切换',
+    tag: 'Tag 标签',
     table: 'Table 表格',
     tooltip: 'Tooltip 文字提示',
     popover: 'Popover 气泡卡片',
@@ -25,45 +26,50 @@ module.exports = {
     confirm: 'confirm 方法组件',
     form: 'Form 表单',
     alert: 'Alert 警告框',
+    // 'test-alert': 'MDX 测试',
     panel: 'Panel 面板',
     collapse: 'Collapse 折叠面板',
     upload: 'Upload 上传',
-    tree: 'Tree 控件',
+    tree: 'Tree 树形控件',
     input: 'Input 输入框',
-    'date-picker': 'Date Picker 日期选择器',
-    'time-picker': 'Time Picker 时间选择器',
+    'date-picker': 'DatePicker 日期选择器',
+    'time-picker': 'TimePicker 时间选择器',
     counter: 'Counter 加减器',
     navMenu: 'NavMenu 导航菜单',
     menu: 'Menu 菜单',
     select: 'Select 选择器',
-    cascader: 'Cascader 级联选择',
+    cascader: 'Cascader 级联选择器',
     radio: 'Radio 单选按钮',
     checkbox: 'Checkbox 复选按钮',
     grid: 'Grid 栅格',
     stepper: 'Stepper 步骤',
     icon: 'Icon 图标',
     progress: 'Progress 进度条',
-    card: 'Card  卡片',
+    card: 'Card 卡片',
     timeline: 'Timeline 时间轴',
     transfer: 'Transfer 穿梭框',
     switch: 'Switch 开关',
     rate: 'Rate 评分'
   },
   designs: {
+    summarize: '概述',
+    'design-values': '设计价值观',
     'design-patterns': '设计模式',
+    'design-principles': '设计原则',
     'about-hiui': '关于 HIUI',
     'visual-framework': '视觉框架',
     start: '快速上手',
-    principles: '设计原则',
-    values: '设计价值观',
     overview: '概览',
-    'layout-grid': '布局/栅格',
-    navigation: '导航/面包屑',
-    copywriting: '文案',
-    colors: '色彩',
+    grid: '栅格',
     layout: '布局',
-    theme: '主题',
-    icon: '图标'
+    'content-layout': '布局',
+    navigation: '导航',
+    copy: '文案',
+    color: '色彩',
+    palette: '配色主题',
+    icon: '图标',
+    font: '字体',
+    image: '图片'
   },
   templates: {
     home: '首页',
@@ -92,6 +98,7 @@ module.exports = {
     home: '首页',
     components: '文档',
     design: '设计语言',
-    template: '模板'
+    template: '模板',
+    github: 'GitHub'
   }
 }
