@@ -1,4 +1,5 @@
 import React from 'react'
+import { render } from 'react-dom'
 import { connect } from 'react-redux'
 import './style/index.scss'
 
@@ -12,45 +13,65 @@ class Component extends React.Component {
       next: '',
       page: '',
       cComponent: null,
-      topNav: 'docs'
+      topNav: 'docs',
+      activeAnchor: ''
     }
-    this.hashChangeEvent = this.hashChangeEvent.bind(this)
     this.contentRef = React.createRef()
   }
-
-  componentDidMount () {
+  componentDidUpdate (prevProps) {
     window.scrollTo(0, 0)
+  }
+  componentDidMount () {
     this.getCurrentPage(() => {
-      this.setState({
-        cComponent: this.getComponent(this.state.page)
-      }, () => {
-        this.getAnchors()
-      })
+      this.setState(
+        {
+          cComponent: this.getComponent(this.state.page)
+        },
+        () => {
+          this.getAnchors()
+          const anchorsDOM = document.querySelectorAll('#markdown-content > div > h2')
+          const anchorsDOMList = [].slice.call(anchorsDOM)
+          anchorsDOMList.map((v, i) => {
+            v.id = v.innerHTML
+
+            render(
+              <span>
+                {v.innerHTML}
+                <a
+                  href={'#' + v.innerHTML}
+                  onClick={() => {
+                    this.setActiveAnchor(v.id)
+                  }}
+                >
+                  <i className='hi-icon icon-maodian' />
+                </a>
+              </span>,
+              v
+            )
+          })
+          if (window.location.hash) {
+            const activeAnchor = decodeURI(window.location.hash.split('#')[1])
+            this.setActiveAnchor(activeAnchor)
+          }
+        }
+      )
     })
 
     this.collectNavs(() => {
       this.getSiblingNav()
     })
   }
-  hashChangeEvent () {
-    let routes = window.location.hash.split('/')
-    if (routes[2] !== 'docs') { return }
-    window.scrollTo(0, 0)
-    this.getCurrentPage()
-    this.getAnchors()
-    this.getSiblingNav()
+  setActiveAnchor = id => {
+    this.setState({ activeAnchor: id })
   }
-
   getAnchors () {
-    const anchorsDOM = document.querySelectorAll('#markdown-content h3')
+    const anchorsDOM = document.querySelectorAll('#markdown-content h2')
     const anchorsDOMList = [].slice.call(anchorsDOM)
     const anchors = anchorsDOMList.map((v, i) => {
-      const id = 'component-anchors-' + i
-      anchorsDOM[i].id = id
-      return {id, text: anchorsDOM[i].innerHTML}
+      return { id: v.id, text: anchorsDOM[i].innerHTML }
     })
 
-    this.setState({anchors})
+    this.setState({ anchors })
   }
 
   getSiblingNav () {
@@ -65,62 +86,56 @@ class Component extends React.Component {
       to: tempArr[index + 1],
       text: footNavs[tempArr[index + 1]] ? footNavs[tempArr[index + 1]] : ''
     }
-    this.setState({pre, next})
+    this.setState({ pre, next })
   }
 
   // 收集所有导航
   collectNavs (fn) {
     let footNavs = []
-    let page = this.props.match.path.split('/')[2]
+    let page = this.props.match.path.split('/')[3]
     footNavs = this.props[page] || {}
-
-    this.setState({footNavs, topNav: page}, fn)
+    this.setState({ footNavs, topNav: page }, fn)
   }
 
   getCurrentPage (fn) {
-    let page = this.props.match.path.split('/')[3]
+    // TODO:这里可能要修改
+    let page = this.props.match.path.split('/')[4]
     page = page || 'quick-start'
     this.setState({ page }, fn)
   }
 
   getComponent (page) {
-    const {
-      theme,
-      locale
-    } = this.props
+    const { theme, locale } = this.props
     // 控制markdown显示隐藏
     const currentPage = this.props.allComponents[this.state.topNav][page]
     if (currentPage) {
-      const el = React.createElement(currentPage.default || currentPage, { theme, locale })
+      const el = React.createElement(currentPage.default || currentPage, {
+        theme,
+        locale
+      })
       return el
     }
   }
 
   render () {
-    const {
-      pre,
-      next,
-      anchors,
-      cComponent,
-      topNav
-    } = this.state
+    const { pre, next, anchors, cComponent, topNav, activeAnchor } = this.state
     return (
       <div className='component'>
         <div className='home-container'>
-          <div className='markdown-content' id='markdown-content'>
+          <div className='markdown-content article' id='markdown-content'>
             {cComponent}
           </div>
 
           <div className='foot-nav clearfix'>
             <a
               className={`pre ${pre.to ? '' : 'none'}`}
-              href={pre.to ? `#/${this.props.locale}/${topNav}/${pre.to}` : ''}
+              href={pre.to ? `/hiui/${this.props.locale}/${topNav}/${pre.to}` : ''}
             >
               {pre.text ? pre.text : ''}
             </a>
             <a
               className={`next ${next.to ? '' : 'none'}`}
-              href={next.to ? `#/${this.props.locale}/${topNav}/${next.to}` : ''}
+              href={next.to ? `/hiui/${this.props.locale}/${topNav}/${next.to}` : ''}
             >
               {next.text ? next.text : ''}
             </a>
@@ -129,23 +144,18 @@ class Component extends React.Component {
 
         <div className='anchor'>
           <ul>
-            {
-              anchors.map((v, i) => (
-                <li key={i}>
-                  <a
-                    onClick={() => {
-                      const target = document.querySelector(`#${v.id}`)
-
-                      if (target) {
-                        target.scrollIntoView({ block: 'start', behavior: 'smooth' })
-                      }
-                    }}
-                  >
-                    {v.text}
-                  </a>
-                </li>
-              ))
-            }
+            {anchors.map((v, i) => (
+              <li key={i} className={activeAnchor === v.text ? 'active' : ''}>
+                <a
+                  href={'#' + v.text}
+                  onClick={() => {
+                    this.setActiveAnchor(v.text)
+                  }}
+                >
+                  {v.text}
+                </a>
+              </li>
+            ))}
           </ul>
         </div>
       </div>
@@ -158,6 +168,7 @@ export default connect(state => ({
   theme: state.global.theme,
   locale: state.global.locale,
   designs: state.global.designNavs,
+  templates: state.global.templatesNavs,
   docs: state.global.componentsNavs,
   allComponents: state.global.components
 }))(Component)
